mcp-server/requirements.txt
mcp[cli]>=1.12.0
python-gitlab>=4.0.0
pydantic>=2.0.0
aiohttp>=3.9.0
python-dotenv>=1.0.0
requests>=2.31.0 